+++ src/store/index.ts
@@ -13,7 +13,7 @@ const persistedReducers = persistCombineReducers<ApplicationState>(
     key: 'YaTA:store',
     migrate: createMigrate(migrations as any, { debug: false }),
     storage,
-    version: 28,
+    version: 29,
     whitelist: ['settings', 'user', 'notes'],
   },
   reducers